refactor: Move `dot` to `render` module
