+++ src/lib.rs
@@ -63,10 +63,10 @@ use pyo3::prelude::*;
 use serde::{Deserialize, Deserializer, Serialize, Serializer};
 
 pub mod algorithms;
-pub mod dot;
 pub mod hierarchy;
 pub mod multiportgraph;
 pub mod portgraph;
+pub mod render;
 pub mod secondary;
 pub mod unmanaged;
 pub mod view;
